Move StaticContainerView
